Refactoring: Extract the upload function.
